requirements.txt
streamlit>=1.28.0
ultralytics==8.0.20
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1+cu118  # Specific version for CUDA compatibility
torchvision==0.15.2+cu118
